requirements.txt
async-timeout==4.0.3
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
dnspython==1.16.0
Flask==3.0.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
load-dotenv==0.1.0
MarkupSafe==2.1.3
mongo==0.2.0
packaging==23.2
pymongo==4.6.1
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
six==1.16.0
spotipy==2.23.0
urllib3==2.0.7
Werkzeug==3.0.1
